Refactor: Move door animation logic to Vehicle.SceneryDoors.cpp
Moves scenery and landscape door functions to a dedicated file
