+++ src/openrct2/libopenrct2.vcxproj
@@ -1130,6 +1130,7 @@
     <ClCompile Include="ride\Vehicle.Collision.cpp" />
     <ClCompile Include="ride\Vehicle.Crash.cpp" />
     <ClCompile Include="ride\Vehicle.Dodgems.cpp" />
+    <ClCompile Include="ride\Vehicle.SceneryDoors.cpp" />
     <ClCompile Include="ride\Vehicle.cpp" />
     <ClCompile Include="ride\Vehicle.Animation.cpp" />
     <ClCompile Include="ride\Vehicle.MiniGolf.cpp">
